separate clan components
