+++ server/controller/controller.js
@@ -97,7 +97,6 @@ module.exports = {
     const db = req.app.get('db'),
     {id} = req.params,
     members = await db.getClanMembers(id)
-    console.log('we in the ctrl.getMem', members)
     res.status(200).send(members);
   },
   top5Users: async (req, res) => {
